+++ lib/src/widgets/helpers/fabulous.dart
@@ -3,15 +3,16 @@
  * See LICENSE for distribution and usage details.
  */
 
-import 'package:flutter/foundation.dart';
-
 import '../../../empathetech_flutter_ui.dart';
 
 import 'dart:io';
 import 'dart:convert';
 import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
+import 'package:flutter/foundation.dart';
 import 'package:file_saver/file_saver.dart';
 import 'package:file_picker/file_picker.dart';
+import 'package:url_launcher/url_launcher.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 
 class EzBackFAB extends FloatingActionButton {
@@ -143,6 +144,114 @@ class EzConfigFAB extends StatelessWidget {
           },
         ),
       ],
+    );
+  }
+}
+
+class EzUpdaterFAB extends StatefulWidget {
+  /// Local app version
+  final String appVersion;
+
+  /// Remote app version (truth)
+  final String versionSource;
+
+  /// Google Play Store URL
+  final String gPlay;
+
+  /// Apple App Store URL
+  final String appStore;
+
+  /// GitHub Releases URL
+  final String github;
+
+  /// [Visibility] wrapped [FloatingActionButton] that links to the latest version if/when there is a mismatch
+  const EzUpdaterFAB({
+    super.key,
+    required this.appVersion,
+    required this.versionSource,
+    required this.gPlay,
+    required this.appStore,
+    required this.github,
+  });
+
+  @override
+  State<EzUpdaterFAB> createState() => _EzUpdaterState();
+}
+
+class _EzUpdaterState extends State<EzUpdaterFAB> {
+  // Define the build data //
+
+  String? latestVersion;
+  String? url;
+
+  bool isLatest = true; // True to start to prevent flickering
+
+  // Define custom functions //
+
+  /// Check for Open UI updates (Desktop only)
+  void checkVersion() async {
+    if (isMobile()) return;
+
+    final http.Response response =
+        await http.get(Uri.parse(widget.versionSource));
+
+    if (response.statusCode != 200) return;
+
+    latestVersion = response.body;
+    if (latestVersion != widget.appVersion && latestVersion != null) {
+      final List<int> latestDigits =
+          latestVersion!.split('.').map(int.parse).toList();
+
+      if (latestDigits.length != 3) return;
+
+      final List<int> appDigits =
+          widget.appVersion.split('.').map(int.parse).toList();
+
+      for (int i = 0; i < latestDigits.length; i++) {
+        if (latestDigits[i] > appDigits[i]) {
+          setState(() => isLatest = false);
+          return;
+        } else if (latestDigits[i] < appDigits[i]) {
+          return;
+        } // if == continue
+      }
+    }
+  }
+
+  // Init //
+
+  @override
+  void initState() {
+    super.initState();
+    checkVersion();
+
+    final TargetPlatform platform = getBasePlatform();
+    switch (platform) {
+      case TargetPlatform.android:
+        url = widget.gPlay;
+      case TargetPlatform.iOS:
+        url = widget.appStore;
+      default:
+        url = widget.github;
+    }
+  }
+
+  // Return the build //
+
+  @override
+  Widget build(BuildContext context) {
+    final ColorScheme colorScheme = Theme.of(context).colorScheme;
+
+    return Visibility(
+      visible: !isLatest,
+      child: FloatingActionButton(
+        heroTag: 'updater_fab',
+        onPressed: () => launchUrl(Uri.parse(url ?? widget.github)),
+        tooltip: ezL10n(context).gUpdates,
+        backgroundColor: colorScheme.secondary,
+        foregroundColor: colorScheme.onSecondary,
+        child: EzIcon(Icons.update),
+      ),
     );
   }
 }
